move star rating near like button
